test(cgroups): refactor tests
