+++ util-cgroups/src/measure/v1.rs
@@ -83,10 +83,7 @@ mod tests {
 
     use tempfile::tempdir;
 
-    use crate::{
-        measure::v1::{V1Collector, V1Stats},
-        CgroupHierarchy, CgroupVersion,
-    };
+    use crate::{measure::v1::V1Collector, CgroupHierarchy, CgroupVersion};
 
     #[test]
     fn test_across_hierarchies() -> anyhow::Result<()> {
@@ -97,7 +94,7 @@ mod tests {
         let mut file1 = File::create(file_path)?;
         writeln!(file1, "{}", value)?;
         // file 2
-        let file_path = root.path().join("memory.stat");
+        let file_path = root.path().join("memory.usage_in_bytes");
         let value: u64 = 19;
         let mut file2 = File::create(file_path)?;
         writeln!(file2, "{}", value)?;
@@ -108,20 +105,17 @@ mod tests {
         let hierarchy2 = CgroupHierarchy::manually_unchecked(root.path(), CgroupVersion::V2, vec!["memory"]);
 
         let res = V1Collector::across_hierarchies("/", &[&hierarchy1, &hierarchy2]);
-        assert!(res.is_ok());
         let mut v1_collector = res.unwrap();
         let mut buf: Vec<u8> = vec![];
 
         assert!(v1_collector.cpuacct_usage.is_some());
-        assert!(v1_collector.memory_stat.is_some());
+        assert!(v1_collector.memory_usage.is_some());
 
         let res = v1_collector.measure(&mut buf);
         assert!(res.is_ok());
         let collector = res.unwrap();
-        assert!(collector.cpuacct_usage.is_some());
-        assert!(collector.memory_stat.is_some());
-        assert_eq!(collector.cpuacct_usage.unwrap(), 15);
-        assert_eq!(collector.memory_stat.unwrap(), 19);
+        assert_eq!(collector.cpuacct_usage, Some(15));
+        assert_eq!(collector.memory_usage, Some(19));
 
         Ok(())
     }
