+++ util-cgroups/src/measure/v2/settings.rs
@@ -305,324 +305,174 @@ mod tests {
     }
 
     #[test]
-    fn test_serialize_bool() {
+    fn test_serialize() {
         let serializer = EnabledKeysSerializer;
-        let true_serialized_res = serializer.serialize_bool(true);
-        assert!(true_serialized_res.is_ok());
-        let true_serialized = true_serialized_res.unwrap();
-        match true_serialized {
-            SerializedValue::Flag(bool) => assert_eq!(bool, true),
-            SerializedValue::EnabledSet(_) => assert!(false),
-        }
+        assert!(matches!(
+            serializer.serialize_bool(true).unwrap(),
+            SerializedValue::Flag(_)
+        ));
+        let serializer = EnabledKeysSerializer;
+        assert!(matches!(
+            serializer.serialize_bool(false).unwrap(),
+            SerializedValue::Flag(_)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let false_serialized_res = serializer.serialize_bool(false);
-        assert!(false_serialized_res.is_ok());
-        let false_serialized = false_serialized_res.unwrap();
-        match false_serialized {
-            SerializedValue::Flag(bool) => assert_eq!(bool, false),
-            SerializedValue::EnabledSet(_) => assert!(false),
-        }
-    }
-
-    #[test]
-    fn test_serialize_i8() {
+        assert!(matches!(
+            serializer.serialize_i8(1 as i8),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i8(1 as i8);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_i8(-1 as i8),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i8(-1 as i8);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_i16() {
+        assert!(matches!(
+            serializer.serialize_i16(1 as i16),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i16(1 as i16);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_i16(-1 as i16),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i16(-1 as i16);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_i32() {
+        assert!(matches!(
+            serializer.serialize_i32(1 as i32),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i32(1 as i32);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_i32(-1 as i32),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i32(-1 as i32);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_i64() {
+        assert!(matches!(
+            serializer.serialize_i64(1 as i64),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i64(1 as i64);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_i64(-1 as i64),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_i64(-1 as i64);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_u8() {
+        assert!(matches!(
+            serializer.serialize_u8(1 as u8),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u8(1 as u8);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_u8(152 as u8),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u8(152 as u8);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_u16() {
+        assert!(matches!(
+            serializer.serialize_u16(1 as u16),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u16(1 as u16);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_u16(152 as u16),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u16(152 as u16);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_u32() {
+        assert!(matches!(
+            serializer.serialize_u32(1 as u32),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u32(1 as u32);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_u32(152 as u32),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u32(152 as u32);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_u64() {
+        assert!(matches!(
+            serializer.serialize_u64(1 as u64),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u64(1 as u64);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_u64(152 as u64),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_u64(152 as u64);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_f32() {
+        assert!(matches!(
+            serializer.serialize_f32(1.0 as f32),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_f32(1.0 as f32);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_f32(-152.58 as f32),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_f32(-152.58 as f32);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_f64() {
+        assert!(matches!(
+            serializer.serialize_f64(-87.0 as f64),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_f64(-87.0 as f64);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_f64(-789.0 as f64),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_f64(-789.0 as f64);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_char() {
+        assert!(matches!(
+            serializer.serialize_char('5'),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_char('5');
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_char('a'),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_char('a');
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_str() {
+        assert!(matches!(
+            serializer.serialize_str("Remus"),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_str("Remus");
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_str("Romulus"),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-        let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_str("Romulus");
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_bytes() {
         let serializer = EnabledKeysSerializer;
         let arr: [u8; 3] = [5, 7, 12];
-        let serialized_res = serializer.serialize_bytes(&arr);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-
+        assert!(matches!(
+            serializer.serialize_bytes(&arr),
+            Err(SerializationError::UnsupportedType)
+        ));
         let serializer = EnabledKeysSerializer;
         let arr: [u8; 0] = [];
-        let serialized_res = serializer.serialize_bytes(&arr);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
-
-    #[test]
-    fn test_serialize_none() {
-        let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_none();
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
+        assert!(matches!(
+            serializer.serialize_bytes(&arr),
+            Err(SerializationError::UnsupportedType)
+        ));
 
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_none();
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        };
-    }
+        assert!(matches!(
+            serializer.serialize_none(),
+            Err(SerializationError::UnsupportedType)
+        ));
+        let serializer = EnabledKeysSerializer;
+        assert!(matches!(
+            serializer.serialize_none(),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_some() {
         #[derive(Serialize)]
         enum People {
             _Claudius,
@@ -633,122 +483,83 @@ mod tests {
             Romulus,
         }
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_some(&People::Romulus);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        }
-    }
+        assert!(matches!(
+            serializer.serialize_some(&People::Romulus),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_unit() {
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_unit();
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        }
-    }
+        assert!(matches!(
+            serializer.serialize_unit(),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_unit_struct() {
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_unit_struct("name");
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        }
-    }
+        assert!(matches!(
+            serializer.serialize_unit_struct("name"),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_unit_variant() {
         let serializer = EnabledKeysSerializer;
-        let serialized_res = serializer.serialize_unit_variant("name", 1 as u32, "variant");
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        }
-    }
+        assert!(matches!(
+            serializer.serialize_unit_variant("name", 1 as u32, "variant"),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_newtype_struct() {
         let serializer = EnabledKeysSerializer;
         let value = 12 as i32;
-        let serialized_res = serializer.serialize_newtype_struct("name", &value);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        }
-    }
+        assert!(matches!(
+            serializer.serialize_newtype_struct("name", &value),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_newtype_variant() {
         let serializer = EnabledKeysSerializer;
         let arr: [i32; 3] = [1, 2, 3];
-        let serialized_res = serializer.serialize_newtype_variant("name", 18 as u32, "variant", &arr);
-        assert!(serialized_res.is_err());
-        let error = serialized_res.expect_err("Expected an error of type SerializationError::UnsupportedType");
-        match error {
-            SerializationError::UnsupportedType => assert!(true),
-            SerializationError::Message(_) => assert!(false),
-        }
-    }
+        assert!(matches!(
+            serializer.serialize_newtype_variant("name", 18 as u32, "variant", &arr),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_seq() {
         let serializer = EnabledKeysSerializer;
         let value = 500 as usize;
-        let serialized_res = serializer.serialize_seq(Some(value));
-        assert!(serialized_res.is_err());
-    }
+        assert!(matches!(
+            serializer.serialize_seq(Some(value)),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_tuple() {
         let serializer = EnabledKeysSerializer;
         let value = 500 as usize;
-        let serialized_res = serializer.serialize_tuple(value);
-        assert!(serialized_res.is_err());
-    }
+        assert!(matches!(
+            serializer.serialize_tuple(value),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_tuple_struct() {
         let serializer = EnabledKeysSerializer;
         let value = 500 as usize;
-        let serialized_res = serializer.serialize_tuple_struct("name", value);
-        assert!(serialized_res.is_err());
-    }
+        assert!(matches!(
+            serializer.serialize_tuple_struct("name", value),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_tuple_variant() {
         let serializer = EnabledKeysSerializer;
         let value = 500 as usize;
-        let serialized_res = serializer.serialize_tuple_variant("name", 14 as u32, "variant", value);
-        assert!(serialized_res.is_err());
-    }
+        assert!(matches!(
+            serializer.serialize_tuple_variant("name", 14 as u32, "variant", value),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_map() {
         let serializer = EnabledKeysSerializer;
         let value = 500 as usize;
-        let serialized_res = serializer.serialize_map(Some(value));
-        assert!(serialized_res.is_err());
-    }
+        assert!(matches!(
+            serializer.serialize_map(Some(value)),
+            Err(SerializationError::UnsupportedType)
+        ));
 
-    #[test]
-    fn test_serialize_struct_variant() {
         let serializer = EnabledKeysSerializer;
         let value = 500 as usize;
-        let serialized_res = serializer.serialize_struct_variant("name", 5 as u32, "variant", value);
-        assert!(serialized_res.is_err());
+        assert!(matches!(
+            serializer.serialize_struct_variant("name", 5 as u32, "variant", value),
+            Err(SerializationError::UnsupportedType)
+        ));
     }
 }
